Safe-global migration
Moved Safe.global from node_modules to sub_modules for ease in scenarios when dealing with multiple contracts that depend on specific versions of external contracts or libraries. This method ensures that all team members and the deployment process use the exact version of a dependency, reducing the risk of inconsistencies. However, it also requires careful management to keep the submodules updated and to handle potential conflicts between different versions of the same dependency. For blockchain projects, where contract immutability is a key feature, having precise control over dependencies is crucial.