Firebase -> Server 0.1.9
Most of the firebase setup is finished, and everything we're doing is now building the serverless server. Thought renaming this to "Server" is a fairer representation, from a task that initially started focused on initiating Firebase Firestore.

Also, with 0.1.8 and this, I've also finished building the skeleton for the whole server\api directory. Only need to write it out.