Smart Contract 0.8.0
Structure has been revamped for clarity. Removed one of the Steez facets for redundancies (merged sip and management), and created a steelo facet for content collections. Fixed most diamond storage references, and moved most constants and variables to the relevant ConstDiamond/LibDiamond. Improved some of the facets, which makes 0.8.x the perfect canvas to build all our features, and aim for 0.9.x to focus on future-proofing and battle-testing before 1.0.x hopefully associated with a closed alpha this summer 2024. Also cleaned up the whole diamond structure, removing now redundant interfaces (which may be needed in the future for 3rd party developers), and renamed all facets for clear separation of tasks. e.g., steezFeesFacet is now simply feesFacet, as Steelo fees are managed by a simple constant. Staking doesn't exist for Steez, so calling it SteeloStaking was also redundant. Oracle was also moved to "App", to fix it's lower importance in the diamond structure. More to come soon in the 0.8 series of smart contract upgrades!